Fix delete styling, TODO fix unnecessary not found render
