+++ src/app/admin/products/[barcode]/ProductView.tsx
@@ -1,28 +1,14 @@
 "use client";
 
 import Barcode from "@/components/Barcode";
-import {
-  AlertDialogFooter,
-  AlertDialogHeader,
-} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import { useToast } from "@/components/ui/use-toast";
-import { Product, deleteProduct } from "@/server/requests/productRequests";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@radix-ui/react-alert-dialog";
+import { Product } from "@/server/requests/productRequests";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
+import ProductDeleteButton from "./ProductDeleteForm";
 
 export const ProductView = ({ product }: { product: Product }) => {
   const { toast } = useToast();
-  const router = useRouter();
 
   return (
     <div className="flex h-full w-full flex-col justify-between gap-y-4">
@@ -80,49 +66,7 @@ export const ProductView = ({ product }: { product: Product }) => {
           <Button asChild>
             <Link href={`/admin/products/${product.barcode}/edit`}>Edit</Link>
           </Button>
-
-          <AlertDialog>
-            <AlertDialogTrigger asChild>
-              <Button variant="destructive">Delete Product</Button>
-            </AlertDialogTrigger>
-            <AlertDialogContent>
-              <AlertDialogHeader>
-                <AlertDialogTitle>Delete Product?</AlertDialogTitle>
-                <AlertDialogDescription>
-                  This will delete the listing for this product. Product data
-                  will remain attached to past purchase history.
-                </AlertDialogDescription>
-              </AlertDialogHeader>
-              <AlertDialogFooter>
-                <AlertDialogCancel>Cancel</AlertDialogCancel>
-                <AlertDialogAction asChild>
-                  <Button
-                    className="bg-red-500 hover:bg-red-600"
-                    onClick={async () => {
-                      try {
-                        const deletedProduct = await deleteProduct(
-                          product.barcode,
-                        );
-                        if (!deletedProduct) {
-                          throw new Error("Product not deleted");
-                        }
-                        toast({ title: "Product deleted", duration: 3000 });
-                        router.replace("/admin/products");
-                      } catch (error) {
-                        router.replace("/admin/products/" + product.barcode);
-                        toast({
-                          title: "Error deleting product",
-                          duration: 3000,
-                        });
-                      }
-                    }}
-                  >
-                    Delete
-                  </Button>
-                </AlertDialogAction>
-              </AlertDialogFooter>
-            </AlertDialogContent>
-          </AlertDialog>
+          <ProductDeleteButton product={product} />
         </div>
       </div>
     </div>
