+++ src/server/requests/productRequests.ts
@@ -1,6 +1,5 @@
 "use server";
 
-import { revalidateTag } from "next/cache";
 import { authenticated } from "../wrappers";
 import { QueryKeys } from "./queryKeys";
 
@@ -162,10 +161,5 @@ export const deleteProduct = (barcode: string) => {
     {
       method: "DELETE",
     },
-  )
-    .then((data) => {
-      revalidateTag(QueryKeys.products);
-      return data;
-    })
-    .then((data) => data.deletedProduct);
+  ).then((data) => data.deletedProduct);
 };
